requirements.txt
django
djangorestframework
django-gtts
django-ckedtor
newspaper3k
